api/src/llm_extraction/requirements.txt
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.64.0
datasets>=2.14.0
